DBZ-3155 Move LogMinerDmlEntry & impl
