+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/parser/DmlParser.java
@@ -5,7 +5,6 @@
  */
 package io.debezium.connector.oracle.logminer.parser;
 
-import io.debezium.connector.oracle.logminer.valueholder.LogMinerDmlEntry;
 import io.debezium.relational.Table;
 
 /**
