+++ debezium-connector-oracle/src/main/java/io/debezium/connector/oracle/logminer/parser/LogMinerDmlEntryImpl.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.connector.oracle.logminer.valueholder;
+package io.debezium.connector.oracle.logminer.parser;
 
 import java.util.Arrays;
 import java.util.Objects;
